+++ clients/admin-ui/src/features/dataset/DataCategoryInput.tsx
@@ -1,18 +1,23 @@
 import {
   Box,
   Button,
+  FormLabel,
+  Grid,
   Menu,
   MenuButton,
   MenuDivider,
   MenuItem,
   MenuList,
   MenuOptionGroup,
+  Stack,
   Text,
 } from "@fidesui/react";
 import { useEffect, useMemo, useState } from "react";
 
 import CheckboxTree from "../common/CheckboxTree";
 import { ArrowDownLineIcon } from "../common/Icon";
+import QuestionTooltip from "../common/QuestionTooltip";
+import DataCategoryTag from "../taxonomy/DataCategoryTag";
 import { transformDataCategoriesToNodes } from "../taxonomy/helpers";
 import { DataCategory } from "../taxonomy/types";
 
@@ -20,13 +25,14 @@ interface Props {
   dataCategories: DataCategory[];
   checked: string[];
   onChecked: (newChecked: string[]) => void;
+  tooltip?: string;
 }
 
 const DataCategoryDropdown = ({
   dataCategories,
   checked,
   onChecked,
-}: Props) => {
+}: Omit<Props, "tooltip">) => {
   const dataCategoryNodes = useMemo(
     () => transformDataCategoriesToNodes(dataCategories),
     [dataCategories]
@@ -111,4 +117,48 @@ const DataCategoryDropdown = ({
   );
 };
 
-export default DataCategoryDropdown;
+const DataCategoryInput = ({
+  dataCategories,
+  checked,
+  onChecked,
+  tooltip,
+}: Props) => {
+  const handleRemoveDataCategory = (dataCategoryName: string) => {
+    onChecked(checked.filter((dc) => dc !== dataCategoryName));
+  };
+
+  const sortedCheckedDataCategories = checked
+    .slice()
+    .sort((a, b) => a.localeCompare(b));
+
+  return (
+    <Grid templateColumns="1fr 3fr">
+      <FormLabel>Data Categories</FormLabel>
+      <Stack>
+        <Box display="flex" alignItems="center">
+          <Box mr="2" width="100%">
+            <DataCategoryDropdown
+              dataCategories={dataCategories}
+              checked={checked}
+              onChecked={onChecked}
+            />
+          </Box>
+          <QuestionTooltip label={tooltip} />
+        </Box>
+        <Stack>
+          {sortedCheckedDataCategories.map((dc) => (
+            <DataCategoryTag
+              key={dc}
+              name={dc}
+              onClose={() => {
+                handleRemoveDataCategory(dc);
+              }}
+            />
+          ))}
+        </Stack>
+      </Stack>
+    </Grid>
+  );
+};
+
+export default DataCategoryInput;